App/requirements.txt
torch
torchvision
torchaudio
llama-index 
llama-index-llms-openai-like
llama-index-llms-huggingface
llama_index-embeddings-huggingface
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
transformers[torch]
huggingface_hub[inference]
streamlit
streamlit-modal
python-dotenv
pymupdf
sentence-transformers
xgboost
uvicorn
fastapi
pandas